✨ feat(tts): Add local player
