+++ src/app/chat/features/Conversation/ChatList/Extras/TTS/InitPlayer.tsx
@@ -1,0 +1,93 @@
+import { memo, useCallback, useEffect, useState } from 'react';
+import { useTranslation } from 'react-i18next';
+
+import { useTTS } from '@/hooks/useTTS';
+import { useChatStore } from '@/store/chat';
+import { filesSelectors, useFileStore } from '@/store/file';
+import { ChatMessageError, ChatTTS } from '@/types/chatMessage';
+import { getMessageError } from '@/utils/fetch';
+
+import Player from './Player';
+
+export interface TTSProps extends ChatTTS {
+  content: string;
+  id: string;
+  loading?: boolean;
+}
+
+const InitPlayer = memo<TTSProps>(({ id, file, voice, content }) => {
+  const [isStart, setIsStart] = useState(false);
+  const [error, setError] = useState<ChatMessageError>();
+  const uploadTTS = useFileStore(filesSelectors.uploadTTSByArrayBuffers);
+  const { t } = useTranslation('chat');
+
+  const [ttsMessage, clearTTS] = useChatStore((s) => [s.ttsMessage, s.clearTTS]);
+
+  const setDefaultError = useCallback(
+    (err?: any) => {
+      setError({ body: err, message: t('tts.responseError', { ns: 'error' }), type: 500 });
+    },
+    [t],
+  );
+
+  const { isGlobalLoading, audio, start, stop, response } = useTTS(content, {
+    onError: (err) => {
+      stop();
+      setDefaultError(err);
+    },
+    onErrorRetry: (err) => {
+      stop();
+      setDefaultError(err);
+    },
+    onSuccess: async () => {
+      if (!response || response.ok) return;
+      const message = await getMessageError(response);
+      if (message) {
+        setError(message);
+      } else {
+        setDefaultError();
+      }
+      stop();
+    },
+    onUpload: async (currentVoice, arrayBuffers) => {
+      if (file && currentVoice === voice) return;
+      const fileID = await uploadTTS(id, arrayBuffers);
+      ttsMessage(id, { file: fileID, voice: currentVoice });
+    },
+  });
+
+  const handleInitStart = useCallback(() => {
+    if (isStart) return;
+    start();
+    setIsStart(true);
+  }, [isStart, start]);
+
+  const handleDelete = useCallback(() => {
+    stop();
+    clearTTS(id);
+  }, [stop, id]);
+
+  const handleRetry = useCallback(() => {
+    setError(undefined);
+    start();
+  }, [start]);
+
+  useEffect(() => {
+    setTimeout(() => {
+      handleInitStart();
+    }, 100);
+  }, []);
+
+  return (
+    <Player
+      audio={audio}
+      error={error}
+      isLoading={isGlobalLoading}
+      onDelete={handleDelete}
+      onInitPlay={handleInitStart}
+      onRetry={handleRetry}
+    />
+  );
+});
+
+export default InitPlayer;
